Fix bug #5082: Outline visibility not remembered.
When a document with a InsetTOC was closed, the outline was closed. But this shouldn't happen because editedInset("toc") does never return something.

The reason that this happened was that if currentBufferView() is 0, we don't even care wehther inset was specified. If inset is specified, it should match the editedInset.

git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@37094 a592a061-630c-0410-9148-cb99ea01b6c8