Fix bug #7261: Assertion when moving up
There are two possibilities to add an Inset to the CoordCache:
- 1), use CoordCache::add(Inset * p, int x, int y),
- 2), use CoordCache::add(Inset * p, Dimension dim).

In the latter case, the x and y are happily initialized as 0.
Also, when one call CoordCache::has(Inset * p), it will happily
return true because the cache knows something about this inset.

In #7261, LyX asked for the (x,y) position of an Inset and
received (0,0) as an answer, although that is not the position
of the Inset.

By adding some defaults for x,y in case of initializing with
dim only will give an extra check whether the cache really
"has" the position object.

In the long run, we might want to split the cache for
dimensions and positions.



git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@37968 a592a061-630c-0410-9148-cb99ea01b6c8