(Replaying r38819 from trunk)
Addressing #7581: now I'm copying the document class as well
from the main document WA to the find/replace WAs (in addition
to the language, as required for addressing #6560).
I'm not cloning the whole BufferParams? because we learned
from

    http://www.mail-archive.com/lyx-devel@lists.lyx.org/msg168175.html

that, for example, if we inherit the default master by mistake,
then we may run into problems.


git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/branches/BRANCH_2_0_X@38845 a592a061-630c-0410-9148-cb99ea01b6c8