Fixup for bug removal in GCC 4.7
A long standing bug has been fixed in GCC 4.7, the bug was a non-
sactioned extra lookup. This caused some code to work that really
shouldn't.

The fixes are: add this->, Class::, or move functions about
as required to fullfill the rules.

In this case some template instantiations
where move to after what they reference.

git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@39813 a592a061-630c-0410-9148-cb99ea01b6c8