Attempt to fix #7673 in a more "interesting" way than in branch (where we will be cautious, of course). The problem was that we were issuing the Buffer::changed() signal before we did updateBuffer(), and this caused an inconsistency. The idea here is to defer issuing this signal until we call processUpdateFlags(). We know we need a redraw if we've deleted a whole paragraph.
This should work properly, so long as checkDepm is called from within
the dispatch mechanism. There may, however, be other paths, and I've
noted one explicitly with a FIXME in Text2.cpp. I've tested a few
different variations, however, and I haven't seen any problems. But if
we do run into problems, we can go ahead and do the update there that we
were previously doing in checkDepm itself.

git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@40352 a592a061-630c-0410-9148-cb99ea01b6c8