branch: Fix bug #7755: Unknown tabular feature "from-dialog"
"from-dialog" is added to the LFUN_INSET_MODIFY function when it is issued from the table settings dialog. This is done to prevent the checking of the individual parameters, because it has to consider all parameters alltogether. Besides, when issued from the dialog it is already guaranteed that the parameters are valid.

This parameter should not be passed onto tabularfeatures.

see r40376.

git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/branches/BRANCH_2_0_X@40653 a592a061-630c-0410-9148-cb99ea01b6c8