Re-encode stmary10 font
Use a 1:1 unicode "encoding" as for all other math symbol fonts.
This is not correct for using the font outside of LyX, but in LyX we misuse
the first 255 code points for symbol fonts (see code points of existing fonts
in lib/symbols). These code points are identical with the postscript versions
of the fonts, so if qt was able to use postscript fonts, they could be used
instead of the truetype versions.
The character varcurlywedge is duplicated at position 254, since qt refuses to
display a font at the tab position (9).
In addition, get rid of the broken private directory which is not needed for
generating ttf.