configure.py: Check path accessibility before call to os.path.isfile.
If you have an unmounted dir, ac_dir, in your PATH, the call to

os.path.isfile( os.path.join(ac_dir, ac_word + ext) )

hangs. This is probably a python bug, but the result of configure.py
hanging and LyX freezing is really bad, hence this workaround.

According to the python docs, MacOS doesn't provide os.access();
the hasattr protection is used for this reason.