tufte-{book,handout}: set default output to pdf2
If the 'nols' option were set, the files could be output with plain
LaTeX as well; but because letterspacing is an important part of the
Tufte design, letterspacing is preserved and pdf2 is set as the
default output format. This is now explained in a note.
(cherry picked from commit 1b7e9178f071a1abdbd6ecea3e60ae28a284b4ec)